requirements.txt
fastapi==0.111.0

uvicorn==0.29.0

python-dotenv==1.0.1

requests==2.31.0

numpy==2.1.0

scikit-learn==1.6.1

sentence-transformers==2.7.0

groq==0.5.0

google-generativeai==0.5.4

tiktoken==0.8.0

streamlit==1.38.0

pandas==2.2.2

pydantic==2.7.1